created a package configuration
