+++ package/__init__.py
@@ -1,0 +1,2 @@
+from .Point import Point
+from .Grid import Grid
